refactor(nx): move the qiibee redux slice into the lib (feature owns its state)
The slice was app-level in store/qiibee/, and the lib reached into it via the
allow-listed store/* alias (a feature->app import). Move it into the lib so the
feature owns its state; the app now just composes the reducer (app->feature —
the correct Nx direction, and the loophole is gone).

- store/qiibee/* -> libs/qiibee/src/lib/state/* (history preserved; internal
  relative imports intact)
- the 5 connected components import ../state/actions instead of store/qiibee/actions
- the barrel exports qiibeeReducer; store/index.ts composes it from
  @my-portfolio/qiibee
- drop the now-unused allowJs + store/* path from the qiibee lib tsconfig
- refresh the qiibee README + CLAUDE.md

store/ now holds only the app-level composition (store/index.ts) + the vestigial
blocpal slice.

Verified: nx run-many typecheck 12/12, root tsc clean, lint 0 errors, build
24/24, tests 3/3 (7) incl. the redux tripwire (makeStore composes the lib reducer).
