+++ CLAUDE.md
@@ -70,7 +70,7 @@ The app's legacy convenience aliases (`components/*`, `store/*`, `util/*`, `imag
 - **App aliases** (`components/*`, `store`, `store/*`, `util/*`, `images/*`) live in **`tsconfig.json` only** — do **not** add them to `tsconfig.base.json` or the boundary rule would flag every in-app import.
 
 ### State management
-- **Redux (qiibee only):** the redux **slice stays app-level** — [store/qiibee/](store/qiibee/) (reducer/actions/dummyData) is composed in [store/index.ts](store/index.ts) via RTK `configureStore` + `next-redux-wrapper` (immutability & serializability checks are **off** because the legacy reducer mutates in place). The `qiibee` **lib**'s `connect()`ed components import `store/qiibee/actions` (allow-listed, so no boundary violation). Redux state is pragmatically typed (`any`). A `blocpal` slice exists but is **not** wired into the store — ignore it.
+- **Redux (qiibee only):** the **qiibee lib owns its slice** — reducer, actions, and the domain types live in [libs/qiibee/src/lib/state/](libs/qiibee/src/lib/state/) and the lib's barrel exports `qiibeeReducer`. The app **composes** it in [store/index.ts](store/index.ts) via RTK `configureStore` + `next-redux-wrapper` — an app→feature dependency, the correct direction (immutability & serializability checks are **off** because the legacy reducer mutates in place). A vestigial `blocpal` slice still sits in `store/` but is **not** wired in — ignore it.
 - **Local:** every other lib uses `useReducer`/`useState`. The assignment reducers (Haptik, Taikai, Fynd, Appbase) keep an untouched `...Copy` of the original list and derive filtered/sorted views from it, so search/sort/reset never lose data.
 
 ### Styling
@@ -100,7 +100,7 @@ The Portfolio landing page ([components/Portfolio/Pages/Home.tsx](components/Por
 
 Durable rules for working in this repo — follow them by default.
 
-- **No `any` — model the real types.** Define interfaces for domain data (e.g. the qiibee `Customer` / `Brand` model in [store/qiibee/types.ts](store/qiibee/types.ts)). For a genuinely dynamic value use `unknown` **plus a narrow typed assertion** (`x as SomeType`) — never `any`. A hand-written redux reducer stays RTK-compatible by typing its action as `{ type: string; data?: unknown }` and casting `data` per case; narrow `currentUser` with `as Customer` / `as Brand` where the action implies the variant. (`unknown` + assertion is fine; `any` is not.)
+- **No `any` — model the real types.** Define interfaces for domain data (e.g. the qiibee `Customer` / `Brand` model in [libs/qiibee/src/lib/state/types.ts](libs/qiibee/src/lib/state/types.ts)). For a genuinely dynamic value use `unknown` **plus a narrow typed assertion** (`x as SomeType`) — never `any`. A hand-written redux reducer stays RTK-compatible by typing its action as `{ type: string; data?: unknown }` and casting `data` per case; narrow `currentUser` with `as Customer` / `as Brand` where the action implies the variant. (`unknown` + assertion is fine; `any` is not.)
 - **Verify before calling it done.** Run the relevant checks and confirm green — `pnpm lint`, `pnpm nx typecheck <lib>` (or `nx run-many -t typecheck`), `pnpm build`, `pnpm testc` — never report "done" on a guess.
 - **Preserve behavior when refactoring.** Type-tightening, renames, and extractions must not change runtime behavior. If a "correct" type would force a behavior change, prefer a documented assertion over changing the logic.
 - **Match the surrounding code.** Keep the existing conventions (the `...Copy` reducer pattern, the `reedeem` misspellings, `arrow-parens: as-needed`, the lib/barrel layout) rather than introducing new styles.
