+++ libs/qiibee/src/index.ts
@@ -5,3 +5,6 @@ export { default as QiibeeBrand } from './lib/Brand';
 export { default as QiibeeCustomer } from './lib/Customer';
 export { default as QiibeeSignIn } from './lib/Login/SignIn';
 export { default as QiibeeSignUp } from './lib/Login/SignUp';
+
+// The redux slice this feature owns; the app composes it in store/index.ts.
+export { default as qiibeeReducer } from './lib/state';
