+++ libs/qiibee/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "//": "Isolated type-check config for the qiibee lib. allowJs + store/* path so the app-level store action imports resolve (as any).",
+  "//": "Isolated type-check config for the qiibee lib.",
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "target": "ES2017",
@@ -12,10 +12,6 @@
     "esModuleInterop": true,
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
-    "allowJs": true,
-    "checkJs": false,
-    "baseUrl": "../..",
-    "paths": { "store/*": ["store/*"] },
     "types": ["node", "react", "react-dom"]
   },
   "include": ["src/**/*.ts", "src/**/*.tsx"]
